modify comment apis

+++ src/services/commentService.ts
@@ -1,8 +1,8 @@
 import { CommentRepo } from '../repos/comment.repo'
-import { UserRepo } from '../repos/user.repo'
 import { CommunityRepo } from '../repos/community.repo'
-import { CommentSchema } from '../utils'
 import APIError from '../errors/APIError'
+import { z } from 'zod'
+import { CommentSchema } from '../utils/zod/commentSchemes'
 
 async function findAllComments(postId: number) {
   const comments = await CommentRepo.findAll(postId)
@@ -16,17 +16,27 @@ async function findComment(postId: number, commentId: number) {
   return comment
 }
 
-async function createComment(data: any) {
-  const validatedData = CommentSchema.parse(data)
-  return await CommentRepo.createComment(validatedData)
+async function createComment(data: z.infer<typeof CommentSchema>) {
+  if (data.parentId) {
+    const commenteExist = await CommentRepo.findComment(
+      data.postId,
+      data.parentId,
+    )
+    if (!commenteExist) throw new APIError('Parent Comment not found', 404)
+  }
+  return await CommentRepo.createComment(data)
 }
 
-async function updateComment(commentId: number, data: any) {
-  const validatedData = CommentSchema.partial().parse(data)
-  return await CommentRepo.updateComment(commentId, validatedData)
+async function updateComment(
+  commentId: number,
+  data: Partial<z.infer<typeof CommentSchema>>,
+) {
+  return await CommentRepo.updateComment(commentId, data)
 }
 
 async function deleteComment(commentId: number) {
+  const comment = await CommentRepo.findCommentById(commentId)
+  if (!comment) throw new APIError('Comment not found', 404)
   return await CommentRepo.deleteComment(commentId)
 }
 
@@ -36,9 +46,6 @@ async function getCommentsByUserIdAndCommunityId(
 ) {
   if (!userId || !communityId)
     throw new APIError('Missing userId or communityId', 400)
-
-  const user = await UserRepo.findById(userId)
-  if (!user) throw new APIError('User not found', 404)
 
   const community = await CommunityRepo.findById(communityId)
   if (!community) throw new APIError('Community not found', 404)
